+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
-## 2.0
+## 2.0.M1
 
 * Migrated from Guice to "bootique-di"
 * #49 Upgrade Logback from v1.1.9 to 1.2.3